fix: moving all to core
As discussed offline move is_nan, is_inf to tract_core
