+++ core/src/ops/math/mod.rs
@@ -710,6 +710,28 @@ element_wise!(sign, Sign, [f16, f32, f64] => |_, xs| {
 };
 q: [i8, u8, i32] => f32::signum);
 
+element_wise_oop!(is_inf, IsInf { detect_positive: bool, detect_negative: bool },
+    [f32] => bool |op, xs, ys| {
+        xs.iter().zip(ys.iter_mut()).for_each(|(x,y)|
+            *y = (op.detect_positive && *x == f32::INFINITY) || (op.detect_negative && *x == f32::NEG_INFINITY)
+        );
+        Ok(())
+    },
+    [f16] => bool |op, xs, ys| {
+        xs.iter().zip(ys.iter_mut()).for_each(|(x,y)|
+            *y = (op.detect_positive && *x == f16::INFINITY) || (op.detect_negative && *x == f16::NEG_INFINITY)
+        );
+        Ok(())
+    }
+);
+
+element_wise_oop!(is_nan, IsNan,
+    [f16, f32] => bool |_, xs, ys| {
+        xs.iter().zip(ys.iter_mut()).for_each(|(x,y)| *y = x.is_nan());
+        Ok(())
+    }
+);
+
 #[cfg(test)]
 mod tests {
     use crate::ops::binary::TypedBinOp;
